workspace_anthropic/b48720cc/requirements.txt
mistralai>=1.0.0
langchain-community>=0.2.0
langchain-core>=0.2.0
chromadb>=0.5.0
sentence-transformers>=2.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
